requirements.txt
tkcalendar>=1.6.1

Pillow>=10.0.0

reportlab>=4.0.0
python-pptx>=0.6.23
